REFACTOR: improve file naming
